extract into FigureSearchExExtras component
